# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "منصة خبرات - Khibrat HR SaaS API",
        "description": "التوثيق الرسمي لكافة واجهات برمجة التطبيقات (APIs) لمنصة خبرات لإدارة الموارد البشرية",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "سيرفر التطوير المحلي"
        }
    ],
    "paths": {
...
        "/api/companies/register": {
            "post": {
                "tags": [
                    "Companies"
                ],
                "summary": "Register a new company and create its initial subscription",
                "operationId": "07492aefe5e7e08d3f7d3c79201cab2e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "address",
                                    "contact_name",
                                    "phone",
                                    "plan_id",
                                    "payment_status"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Acme Company"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "hr@acme.com"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Damascus"
                                    },
                                    "contact_name": {
                                        "type": "string",
                                        "example": "Ahmad"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+963999999999"
                                    },
                                    "plan_id": {
                                        "type": "string",
                                        "example": "00000000-0000-0000-0000-000000000000"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Company registered successfully"
                    }
                }
            }
        },
        "/api/companies/{company}/attendance-policy": {
            "put": {
                "tags": [
                    "Companies"
                ],
                "summary": "تحديث سياسة الحضور والإعدادات الجغرافية للبصمة",
                "description": "يقوم هذا الـ API بحفظ سياسات أوقات العمل للشركة بالإضافة إلى إحداثيات المقر الجغرافي ونصف قطر نطاق التواجد المسموح به للموظفين لتسجيل الحضور.",
                "operationId": "6128e14a9c1d912a118d16e7bea2d378",
                "parameters": [
                    {
                        "name": "company",
                        "in": "path",
                        "description": "معرف الشركة",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "work_start_time",
                                    "work_end_time",
                                    "allowed_late_minutes",
                                    "allowed_early_leave_minutes",
                                    "work_days",
                                    "minimum_daily_hours",
                                    "allows_overtime",
                                    "allowed_perimeter",
                                    "latitude",
                                    "longitude"
                                ],
                                "properties": {
                                    "work_start_time": {
                                        "type": "string",
                                        "example": "08:00:00"
                                    },
                                    "work_end_time": {
                                        "type": "string",
                                        "example": "17:00:00"
                                    },
                                    "allowed_late_minutes": {
                                        "type": "integer",
                                        "example": 15
                                    },
                                    "allowed_early_leave_minutes": {
                                        "type": "integer",
                                        "example": 15
                                    },
                                    "work_days": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "example": "monday"
                                        }
                                    },
                                    "minimum_daily_hours": {
                                        "type": "integer",
                                        "example": 8
                                    },
                                    "allows_overtime": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "allowed_perimeter": {
                                        "type": "integer",
                                        "example": 150
                                    },
                                    "latitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 33.5138
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 36.2765
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Attendance policy updated successfully"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "أدخلي الـ Token الذي حصلتِ عليه من الـ Login هنا لتفعيل الصلاحيات",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}